+++ ft_memset.c
@@ -3,14 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memset.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: relouren <relouren@student.42.com>         +#+  +:+       +#+        */
+/*   By: relouren <relouren@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/16 12:54:02 by relouren          #+#    #+#             */
-/*   Updated: 2022/11/16 12:54:02 by relouren         ###   ########.fr       */
+/*   Updated: 2022/11/16 16:31:23 by relouren         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *s, int c, unsigned int n)
+#include "libft.h"
+
+void	*ft_memset(void *s, int c, size_t n)
 {
 	unsigned char	*ptr;
